S104 split: extract state_helpers.py and ngfw_runtime_ops.py
Begin breaking up engine/provisioner/main.py (3,439 lines, Sonar S104)
into focused sibling modules. main.py keeps re-imports so
patch("main.X") test mocks across the existing 718 provisioner tests
continue to work without test changes.

- state_helpers.py: validation + provider metadata + state builders
  (pure helpers, no I/O beyond os.environ.get)
- ngfw_runtime_ops.py: NGFW start/stop runtime operations. Cross-module
  callees historically patched on main (update_instance_state,
  publish_ngfw_event, get_ngfw_data_by_request_id, AWSExecutor,
  OpsOrchestrator) are accessed via lazy import main; main.X(...).
- main.py keeps the publish_ngfw_event and OpsOrchestrator imports
  with a noqa: F401 marker so the ngfw_runtime_ops lazy-lookup path
  resolves at runtime and the existing test patches keep working.

main.py: 3,439 -> 3,131 lines. Further splits to follow in the same
PR until the file is under the Sonar S104 threshold.
